+++ src/main/java/com/clouway/core/User.java
@@ -1,5 +1,7 @@
 package com.clouway.core;
 
+import java.math.BigDecimal;
+
 /**
  * Created byivan.genchev1989@gmail.com.
  */
@@ -7,17 +9,17 @@ public class User {
 
   public final String name;
   public final String password;
-  private Double amount;
+  private BigDecimal amount;
   public final Session session;
 
-  public User(String name, String password, Double amount, Session session) {
+  public User(String name, String password, BigDecimal amount, Session session) {
     this.name = name;
     this.password = password;
     this.amount = amount;
     this.session = session;
   }
 
-  public Double getAmount() {
+  public BigDecimal getAmount() {
     return amount;
   }
 
@@ -25,11 +27,14 @@ public class User {
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
+
     User user = (User) o;
+
     if (amount != null ? !amount.equals(user.amount) : user.amount != null) return false;
     if (name != null ? !name.equals(user.name) : user.name != null) return false;
     if (password != null ? !password.equals(user.password) : user.password != null) return false;
     if (session != null ? !session.equals(user.session) : user.session != null) return false;
+
     return true;
   }
